pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bee-operation-data"
version = "0.1.0"
description = "Operational metrics and OKR collector for BeeMessage dashboard."
requires-python = ">=3.11"
dependencies = [
  "requests",
  "python-dotenv",
  "supabase",
]

[tool.setuptools]
packages = ["bee_operation_data", "bee_operation_data.http"]
